# mobile/src/api/openapi.json
{
  "components": {
    "schemas": {
...
      "AreaOut": {
        "properties": {
          "planning_area": {
            "title": "Planning Area",
            "type": "string"
          },
          "region": {
            "title": "Region",
            "type": "string"
          }
        },
        "required": [
          "planning_area",
          "region"
        ],
        "title": "AreaOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "SG Events API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/areas/resolve": {
      "get": {
        "operationId": "resolve_postal_areas_resolve_get",
        "parameters": [
          {
            "description": "6-digit Singapore postal code",
            "in": "query",
            "name": "postal",
            "required": true,
            "schema": {
              "description": "6-digit Singapore postal code",
              "title": "Postal",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AreaOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Resolve Postal",
        "tags": [
          "areas"
        ]
      }
    },
...
  }
}